requirements.txt
yfinance
pandas
langchain
langchain-community
langchain-nvidia-ai-endpoints
langchain-core
langgraph
python-dotenv
streamlit
plotly
scipy
numpy
requests
beautifulsoup4
urllib3<2.0
akshare
